Move helpText to CommandUIOptParse
